Shifted the sync pulse forward by one clock relative to the data window in both the clump and the unclump library blocks.
Shifted the sync pulse forward by an additional clock before
entering the XAUI, but then shifted the data forward by one
clock after exiting the XAUI.

This is tested in hardware, and is equivalent to how it was
before, but the logic feels more "correct".

Added a posedge detector to the sync pulse received over the
XAUI out-of-band line.